checkout: switch to pointer for all url.URL since gob seems to be buggy on value

+++ checkout/infrastructure/locker/redis_test.go
@@ -27,11 +27,11 @@ func startUp(t *testing.T) *tempredis.Server {
 
 func getRedisLocker(network, address string) *locker.Redis {
 	redis := locker.NewRedis(&struct {
-		MaxIdle                 float64 `inject:"config:commerce.checkout.placeorder.lock.redis.maxIdle"`
-		IdleTimeoutMilliseconds float64 `inject:"config:commerce.checkout.placeorder.lock.redis.idleTimeoutMilliseconds"`
-		Network                 string  `inject:"config:commerce.checkout.placeorder.lock.redis.network"`
-		Address                 string  `inject:"config:commerce.checkout.placeorder.lock.redis.address"`
-		Database                float64 `inject:"config:commerce.checkout.placeorder.lock.redis.database"`
+		MaxIdle                 int    `inject:"config:commerce.checkout.placeorder.lock.redis.maxIdle"`
+		IdleTimeoutMilliseconds int    `inject:"config:commerce.checkout.placeorder.lock.redis.idleTimeoutMilliseconds"`
+		Network                 string `inject:"config:commerce.checkout.placeorder.lock.redis.network"`
+		Address                 string `inject:"config:commerce.checkout.placeorder.lock.redis.address"`
+		Database                int    `inject:"config:commerce.checkout.placeorder.lock.redis.database"`
 	}{MaxIdle: 3, IdleTimeoutMilliseconds: 240000, Network: network, Address: address, Database: 0})
 	return redis
 }
